Build/DjangoCoreUITemplate/app/requirements.txt
django
psycopg
python-dotenv
django-debug-toolbar # for development
#celery
#django-celery-beat
#redis
#requests
#selenium 
#wget
#docker
#whitenoise
#django-widget-tweaks
#django-admin-interface
#django-extensions
#gunicorn